+++ src/components/Elements/index.js
@@ -9,3 +9,4 @@ export { default as Panels } from './Panels';
 export { default as Panel } from './Panel';
 export { default as DropZone } from './DropZone';
 export { default as NodeBin } from './NodeBin';
+export { default as Form } from './Form';
