+++ src/containers/Elements/Form.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { reduxForm } from 'redux-form';
 import { fromPairs, map } from 'lodash';
-import { Field } from './';
+import { Form } from '../../components/Elements';
 
 const typeInitalValue = (field) => {
   switch (field.type) {
@@ -18,34 +18,22 @@ const typeInitalValue = (field) => {
 const initialValues = fields =>
   fromPairs(map(fields, field => [field.name, typeInitalValue(field)]));
 
-const Form = ({ fields, handleSubmit }) => (
-  <form onSubmit={handleSubmit}>
-    { fields.map((field, index) => (
-      <Field key={index} {...field} />
-    )) }
-    <br />
-    <button type="submit">Submit</button>
-  </form>
-);
-
-Form.propTypes = {
-  fields: PropTypes.array.isRequired,
-  handleSubmit: PropTypes.func.isRequired,
-};
-
-const ReduxedForm = reduxForm({
-  destroyOnUnmount: true,
-  forceUnregisterOnUnmount: true,
-})(Form);
-
 /**
   * Renders a redux form that contains fields according to a `fields` config.
   */
-const AutoInitialisedForm = props =>
-  <ReduxedForm initialValues={initialValues(props.fields)} {...props} />;
+const autoInitialisedForm = (WrappedComponent) => {
+  const AutoInitialisedForm = props => (
+    <WrappedComponent initialValues={initialValues(props.fields)} {...props} />
+  );
 
-AutoInitialisedForm.propTypes = {
-  fields: PropTypes.array.isRequired,
+  AutoInitialisedForm.propTypes = {
+    fields: PropTypes.array.isRequired,
+  };
+
+  return AutoInitialisedForm;
 };
 
-export default AutoInitialisedForm;
+export default autoInitialisedForm(reduxForm({
+  destroyOnUnmount: true,
+  forceUnregisterOnUnmount: true,
+})(Form));
